+++ karma.conf.js
@@ -13,22 +13,24 @@ module.exports = function(config) {
         }
     },
     files: [
-      {pattern: 'node_modules/es6-shim/es6-shim.js', included: true, watched: true},
-      {pattern: 'node_modules/zone.js/dist/zone.js', included: true, watched: true},
-      //Needed for async testsing {https://github.com/angular/angular/issues/8232}
-      {pattern: 'node_modules/zone.js/dist/async-test.js', included: true, watched: true},
-      {pattern: 'node_modules/reflect-metadata/Reflect.js', included: true, watched: true},
+      {pattern: 'node_modules/core-js/client/core.js', included: true, watched: true},
+      'node_modules/reflect-metadata/Reflect.js',
+      'node_modules/zone.js/dist/zone.js',
+      'node_modules/zone.js/dist/long-stack-trace-zone.js',
+      'node_modules/zone.js/dist/proxy.js',
+      'node_modules/zone.js/dist/sync-test.js',
+      'node_modules/zone.js/dist/jasmine-patch.js',
+      'node_modules/zone.js/dist/async-test.js',
+      'node_modules/zone.js/dist/fake-async-test.js',
       {pattern: 'node_modules/systemjs/dist/system-polyfills.js', included: true, watched: false },
       {pattern: 'node_modules/systemjs/dist/system.src.js', included: true, watched: true},
       
       {pattern: 'karma-systemjs-config.js', included: true, watched: true},
       
-      {pattern: 'node_modules/rxjs/**/*', included: false, watched: false},
-      {pattern: 'node_modules/@angular/common/**/*', included: false, watched: false},
-      {pattern: 'node_modules/@angular/core/**/*', included: false, watched: false},
-      {pattern: 'node_modules/@angular/platform-browser/**/*', included: false, watched: false},
-      {pattern: 'node_modules/@angular/compiler/**/*', included: false, watched: false},
-      {pattern: 'node_modules/@angular/platform-browser-dynamic/**/*', included: false, watched: false},
+      {pattern: 'node_modules/rxjs/**/*.js', included: false, watched: false},
+      {pattern: 'node_modules/rxjs/**/*.js.map', included: false, watched: false},
+      {pattern: 'node_modules/@angular/**/*.js', included: false, watched: false},
+      {pattern: 'node_modules/@angular/**/*.js.map', included: false, watched: false},
       
       {pattern: 'karma-test-shim.js', included: true, watched: true},
 
@@ -38,8 +40,8 @@ module.exports = function(config) {
 
       // paths loaded via Angular's component compiler
       // (these paths need to be rewritten, see proxies section)
-      {pattern: 'dist/**/*.html', included: false, watched: true},
-      {pattern: 'dist/**/*.css', included: false, watched: true},
+      // {pattern: 'dist/**/*.html', included: false, watched: true},
+      // {pattern: 'dist/**/*.css', included: false, watched: true},
 
       // paths to support debugging with source maps in dev tools
       {pattern: 'dist/**/*.ts', included: false, watched: false},