+++ src/core/services/managers/marker-manager.ts
@@ -49,12 +49,28 @@ export class MarkerManager {
     return this._markers.get(marker).then((m: Marker) => { m.setIcon(marker.getIcon()) });
   }
 
+  updateOpacity(marker: SebmGoogleMapMarker): Promise<void> {
+    return this._markers.get(marker).then((m: Marker) => m.setOpacity(marker.opacity));
+  }
+
+  updateVisible(marker: SebmGoogleMapMarker): Promise<void> {
+    return this._markers.get(marker).then((m: Marker) => m.setVisible(marker.visible));
+  }
+
+  updateZIndex(marker: SebmGoogleMapMarker): Promise<void> {
+    return this._markers.get(marker).then((m: Marker) => m.setZIndex(marker.zIndex));
+  }
+
   addMarker(marker: SebmGoogleMapMarker) {
     const markerPromise = this._mapsWrapper.createMarker({
       position: { lat: marker.latitude, lng: marker.longitude },
       label: marker.label,
       draggable: marker.draggable,
-      icon: marker.getIcon()
+      icon: marker.getIcon(),
+      opacity: marker.opacity,
+      visible: marker.visible,
+      zIndex: marker.zIndex,
+      title: marker.title
     });
     this._markers.set(marker, markerPromise);
   }